(#18): Rename module `sensorslib` to `sensorlib`
This commit renames the module name from `sensorslib` to `sensorlib`.
