+++ .idea/gradle.xml
@@ -11,7 +11,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/buildSrc" />
-            <option value="$PROJECT_DIR$/sensorslib" />
+            <option value="$PROJECT_DIR$/sensorlib" />
             <option value="$PROJECT_DIR$/wear" />
           </set>
         </option>
