+++ sensorlib/build.gradle
