+++ sensorlib/proguard-rules.pro
