+++ sensorlib/src/main/AndroidManifest.xml
