+++ sensorlib/src/main/java/com/handmonitor/sensorlib/SensorWindowFlow.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorslib
+package com.handmonitor.sensorlib
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.awaitClose
