+++ sensorlib/src/test/java/com/handmonitor/sensorlib/SensorWindowBufferTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorslib
+package com.handmonitor.sensorlib
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.junit5.MockKExtension
