+++ settings.gradle
@@ -15,4 +15,4 @@ dependencyResolutionManagement {
 
 rootProject.name = "HandMonitor"
 include ':wear'
-include ':sensorslib'
+include ':sensorlib'
